refactor: simplify cache system and clean up debug logging
- Simplify cache system to use only localStorage for client-side caching
- Remove complex dual-caching system (localStorage + serverCache)
- Remove debug logging from cache functions after confirming they work
- Add LEADERBOARD_DATA cache duration constant
- Clean up unused imports and improve code consistency

This simplifies the caching architecture while maintaining the same functionality.
The cache system is now more maintainable and easier to understand.
