+++ app/services/leaderboardService.ts
@@ -1,5 +1,5 @@
 import type { LeaderboardEntry } from "./types";
-import { BOOST_CONFIG } from "@/lib/constants";
+import { BOOST_CONFIG, PROJECT_ACCOUNTS_TO_EXCLUDE } from "@/lib/constants";
 
 export interface LeaderboardResponse {
   entries: LeaderboardEntry[];
@@ -8,30 +8,148 @@ export interface LeaderboardResponse {
   nextUpdate?: string | null;
 }
 
-export async function getLeaderboardCreators(
-  page: number = 1,
-  perPage: number = 200,
-): Promise<LeaderboardResponse> {
-  const response = await fetch(
-    `/api/leaderboard/basic?page=${page}&per_page=${perPage}`,
-    {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    },
-  );
+type Profile = {
+  id: string;
+  display_name?: string;
+  name?: string;
+  image_url?: string;
+  scores?: Array<{
+    slug: string;
+    points?: number;
+  }>;
+};
 
-  if (!response.ok) {
-    throw new Error("Failed to fetch leaderboard data");
+/**
+ * Fetch top 200 entries from Talent Protocol API
+ */
+export async function fetchTop200Entries(apiKey: string): Promise<Profile[]> {
+  const baseUrl = "https://api.talentprotocol.com/search/advanced/profiles";
+  const batchSize = 200 + PROJECT_ACCOUNTS_TO_EXCLUDE.length;
+  const totalNeeded = 200 + PROJECT_ACCOUNTS_TO_EXCLUDE.length;
+  let allProfiles: Profile[] = [];
+
+  for (let page = 1; allProfiles.length < totalNeeded; page++) {
+    const data = {
+      query: {
+        score: {
+          min: 1,
+          scorer: "Creator Score",
+        },
+      },
+      sort: {
+        score: { order: "desc", scorer: "Creator Score" },
+        id: { order: "desc" },
+      },
+      page,
+      per_page: batchSize,
+    };
+
+    const queryString = [
+      `query=${encodeURIComponent(JSON.stringify(data.query))}`,
+      `sort=${encodeURIComponent(JSON.stringify(data.sort))}`,
+      `page=${page}`,
+      `per_page=${batchSize}`,
+      `view=scores_minimal`,
+    ].join("&");
+
+    const res = await fetch(`${baseUrl}?${queryString}`, {
+      headers: {
+        Accept: "application/json",
+        "X-API-KEY": apiKey,
+      },
+    });
+
+    if (!res.ok) {
+      const errorText = await res.text();
+      throw new Error(`Failed to fetch page ${page}: ${errorText}`);
+    }
+
+    const json = await res.json();
+    const profiles = json.profiles || [];
+    allProfiles = [...allProfiles, ...profiles];
+
+    if (profiles.length < batchSize) break;
+
+    if (page < Math.ceil(totalNeeded / batchSize)) {
+      await new Promise((resolve) => setTimeout(resolve, 100));
+    }
   }
 
-  const json = await response.json();
+  return allProfiles.slice(0, totalNeeded);
+}
+
+/**
+ * Get top 200 leaderboard entries with boosted profiles integration
+ */
+export async function getTop200LeaderboardEntries(): Promise<LeaderboardResponse> {
+  const apiKey = process.env.TALENT_API_KEY;
+  if (!apiKey) {
+    throw new Error("Missing Talent API key");
+  }
+
+  // Fetch profiles
+  const profiles = await fetchTop200Entries(apiKey);
+
+  // Filter out project accounts
+  const filteredProfiles = profiles.filter(
+    (profile) => !PROJECT_ACCOUNTS_TO_EXCLUDE.includes(profile.id),
+  );
+
+  // Fetch boosted profiles for integration
+  let boostedProfileIds: string[] = [];
+  try {
+    boostedProfileIds = await getBoostedProfilesData();
+  } catch {
+    boostedProfileIds = [];
+  }
+
+  // Map to basic entries (no rewards) with boosted status
+  const mapped = filteredProfiles.map((profile: Profile) => {
+    const creatorScores = Array.isArray(profile.scores)
+      ? profile.scores
+          .filter((s) => s.slug === "creator_score")
+          .map((s) => s.points ?? 0)
+      : [];
+    const score = creatorScores.length > 0 ? Math.max(...creatorScores) : 0;
+    const isBoosted = boostedProfileIds.includes(profile.id);
+
+    return {
+      name: profile.display_name || profile.name || "Unknown",
+      pfp: profile.image_url || undefined,
+      score,
+      id: profile.id,
+      talent_protocol_id: profile.id,
+      isBoosted,
+    };
+  });
+
+  // Sort by score and assign ranks
+  mapped.sort((a, b) => b.score - a.score);
+
+  let lastScore: number | null = null;
+  let lastRank = 0;
+  let ties = 0;
+  const ranked = mapped.map((entry, idx) => {
+    let rank;
+    if (entry.score === lastScore) {
+      rank = lastRank;
+      ties++;
+    } else {
+      rank = idx + 1;
+      if (ties > 0) rank = lastRank + ties;
+      lastScore = entry.score;
+      lastRank = rank;
+      ties = 1;
+    }
+    return { ...entry, rank };
+  });
+
+  // Count boosted creators
+  const boostedCreatorsCount = ranked.filter((entry) => entry.isBoosted).length;
+
   return {
-    entries: json.entries || [],
-    boostedCreatorsCount: json.boostedCreatorsCount,
-    lastUpdated: json.lastUpdated,
-    nextUpdate: json.nextUpdate,
+    entries: ranked,
+    boostedCreatorsCount,
   };
 }
